% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhfCalcGrowthRates.R
\name{rhfCalcGrowth}
\alias{rhfCalcGrowth}
\title{Calcution of Growth Rates}
\usage{
rhfCalcGrowth(x, lag = 12)
}
\arguments{
\item{x}{data}

\item{lag}{time lag}
}
\value{
datatable
}
\description{
This function calculates growth rates on a vector/dataframe column given a time lag.
}
\examples{
\dontrun{
rhfCalcGrowth(x = 1:100, lag = 10)
}

}
