% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhfDeleteDataDB.R
\name{rhfDeleteDB}
\alias{rhfDeleteDB}
\title{Delete databse entries}
\usage{
rhfDeleteDB(con, keys, dbTable)
}
\arguments{
\item{con}{database connection object}

\item{keys}{key column}

\item{dbTable}{database table}
}
\value{
datatable
}
\description{
This function deletes database entries/rows by key.
}
\examples{
\dontrun{
rhfDeleteDB(con = con, keys = keys, dbTable = "databaseTable")
}

}
