% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhfFacToChar.R
\name{rhfFacToChar}
\alias{rhfFacToChar}
\title{Converting factor into character variables}
\usage{
rhfFacToChar(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
dataframe
}
\description{
This function converts factor into character variables.
}
\examples{
\dontrun{
dt <- rhfFacToChar(data = dt)
}

}
