% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhfInfNaToZero.R
\name{rhfInfNAToZero}
\alias{rhfInfNAToZero}
\title{Converts Inf and NAs to 0}
\usage{
rhfInfNAToZero(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
datatable
}
\description{
This function converts all Inf values and NAs to zero.
}
\examples{
\dontrun{
dt <- rhfInfNAToZero(data = dt)
}

}
