% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhfInsertDataDB.R
\name{rhfInsertDataDB}
\alias{rhfInsertDataDB}
\title{Insert dataframe to database}
\usage{
rhfInsertDataDB(data, dbTable, con)
}
\arguments{
\item{data}{dataframe}

\item{dbTable}{database table}

\item{con}{database connection object}
}
\value{
datatable
}
\description{
This function inserts a dataframe into a database.
}
\examples{
\dontrun{
rhfInsertDataDB(data = dt, dbTable = "databaseTable")
}

}
