% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhfPwdInput.R
\name{rhfPwdInput}
\alias{rhfPwdInput}
\title{Password User Input Prompt Dialogue}
\usage{
rhfPwdInput()
}
\value{
user input as character
}
\description{
This function starts a promt dialoque that asks for a password input.
}
\examples{
\dontrun{
pwd <- rhfPwdInput()
}

}
