% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhfRepeatUntilOk.R
\name{rhfRepeatUntilOk}
\alias{rhfRepeatUntilOk}
\title{Repeats Target Function until Result succesful}
\usage{
rhfRepeatUntilOk(expr, silent = TRUE, maxIter = 1000,
  quotedExpr = substitute(expr), envir = parent.frame())
}
\arguments{
\item{expr}{target function expression}

\item{silent}{logical: should the report of error messages be suppressed?}

\item{maxIter}{maximum iterations of target function}

\item{quotedExpr}{substitute(expr) - returns the parse tree for the (unevaluated) expression}

\item{envir}{environment - evaluate an R expression in a specified environment}
}
\value{
result target function
}
\description{
This function repeats the target function until the execution of the target function is successfull.
}
\examples{
\dontrun{
rhfRepeatUntilOk(data <- targetFunction())
}

}
